replace TeamDrawer with AppDrawer
- move AppBanner text to AppBar
- move links from AppBar to AppDrawer
- recolor bluegrey buttons to default
