+++ layouts/default.vue
@@ -1,16 +1,14 @@
 <template>
   <v-app>
-    <no-ssr>
-      <app-bar v-if="$store.getters.authenticated" />
+    <no-ssr v-if="authenticated">
+      <app-bar />
 
-      <team-drawer v-if="teamPage" />
+      <app-drawer />
     </no-ssr>
 
     <app-broadcaster />
 
     <v-content>
-      <app-banner />
-
       <template v-if="teamPage">
         <team-channel />
         <nuxt />
@@ -24,24 +22,24 @@
 
 <script>
   import { Vue, Component, Watch } from 'vue-property-decorator'
+  import { mapGetters } from 'vuex'
   import { Team } from '@/models'
   import AppBar from '@/components/App/Bar'
-  import AppBanner from '@/components/App/Banner'
   import AppBroadcaster from '@/components/App/Broadcaster'
   import TeamChannel from '@/components/Team/Channel'
-  import TeamDrawer from '@/components/Team/Drawer'
-  // import AppDrawer from '@/components/App/Drawer'
+  import AppDrawer from '@/components/App/Drawer'
 
   @Component({
     name: 'App',
     components: {
       AppBar,
-      AppBanner,
       AppBroadcaster,
       TeamChannel,
-      TeamDrawer
-      // AppDrawer
-    }
+      AppDrawer
+    },
+    computed: mapGetters([
+      'authenticated'
+    ])
   })
   export default class Layout extends Vue {
     head () {
