+++ nuxt.config.js
@@ -6,7 +6,7 @@ export default {
 
   server: {
     host: '0.0.0.0',
-    port: '8080'
+    port: process.env.NODE_ENV === 'production' ? '8080' : '4000'
   },
 
   /*
