+++ pages/teams/_teamId/seasons/index.vue
@@ -4,11 +4,7 @@
       <v-col cols="12">
         <competition-form>
           <template #default="{ on }">
-            <v-btn
-              color="blue-grey"
-              dark
-              v-on="on"
-            >
+            <v-btn v-on="on">
               <v-icon left>mdi-plus-circle-outline</v-icon>
               Competition
             </v-btn>
